Fold Address entity into Contact entity
