+++ src/Enfo.Domain/Data/Counties.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+// ReSharper disable StringLiteralTypo
 
 namespace Enfo.Domain.Data
 {
