+++ src/Enfo.Domain/Entities/EpdContact.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using Enfo.Domain.Entities.Base;
 
 namespace Enfo.Domain.Entities
@@ -17,14 +18,31 @@ namespace Enfo.Domain.Entities
         [StringLength(100)]
         public string Organization { get; set; }
 
-        [Required]
-        public Address Address { get; set; }
-        public int? AddressId { get; set; }
-
         [StringLength(50)]
         public string Telephone { get; set; }
 
         [StringLength(100)]
         public string Email { get; set; }
+        
+        // Postal (mailable) addresses only
+
+        [Required]
+        [StringLength(100)]
+        public string Street { get; set; }
+
+        [StringLength(100)]
+        public string Street2 { get; set; }
+
+        [Required]
+        [StringLength(50)]
+        public string City { get; set; }
+
+        [Required]
+        [DefaultValue("GA")]
+        [StringLength(2)]
+        public string State { get; set; } = "GA";
+
+        [StringLength(10)]
+        public string PostalCode { get; set; }
     }
 }
