+++ src/Enfo.Domain/Resources/EpdContact/EpdContactView.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel;
-using Enfo.Domain.Resources.Address;
 using Enfo.Domain.Utils;
 using JetBrains.Annotations;
 
@@ -7,7 +6,7 @@ namespace Enfo.Domain.Resources.EpdContact
 {
     public class EpdContactView
     {
-        public EpdContactView([NotNull] Domain.Entities.EpdContact item)
+        public EpdContactView([NotNull] Entities.EpdContact item)
         {
             Guard.NotNull(item, nameof(item));
 
@@ -16,9 +15,13 @@ namespace Enfo.Domain.Resources.EpdContact
             ContactName = item.ContactName;
             Title = item.Title;
             Organization = item.Organization;
-            Address = new AddressView(item.Address);
             Telephone = item.Telephone;
             Email = item.Email;
+            City = item.City;
+            PostalCode = item.PostalCode;
+            State = item.State;
+            Street = item.Street;
+            Street2 = item.Street2;
         }
 
         public int Id { get; }
@@ -29,8 +32,29 @@ namespace Enfo.Domain.Resources.EpdContact
 
         public string Title { get; }
         public string Organization { get; }
-        public AddressView Address { get; }
         public string Telephone { get; }
         public string Email { get; }
+
+        [DisplayName("Street Address")]
+        public string Street { get; }
+
+        [DisplayName("Apt / Suite / Other")]
+        public string Street2 { get; }
+
+        public string City { get; }
+        public string State { get; }
+
+        [DisplayName("Postal Code")]
+        public string PostalCode { get; }
+        
+        public string AsLinearString
+        {
+            get
+            {
+                var cityState = new[] {City, State}.ConcatNonEmptyStrings(", ");
+                var cityStateZip = new[] {cityState, PostalCode}.ConcatNonEmptyStrings(" ");
+                return new[] {ContactName, Street, Street2, cityStateZip}.ConcatNonEmptyStrings(", ");
+            }
+        }
     }
 }
