+++ src/Enfo.WebApp/Pages/Admin/Delete.cshtml
@@ -56,7 +56,7 @@
                 <td>
                     @if (Model.Item.CommentContact != null)
                     {
-                        @Html.DisplayFor(m => m.Item.CommentContact, "Contact")
+                        @Html.DisplayFor(m => m.Item.CommentContact, "EpdContactView")
                     }
                     else
                     {
@@ -100,7 +100,7 @@
                 </p>
                 @if (Model.Item.HearingContact != null)
                 {
-                    @Html.DisplayFor(m => m.Item.HearingContact, "Contact")
+                    @Html.DisplayFor(m => m.Item.HearingContact, "EpdContactView")
                 }
                 else
                 {
