--- src/Enfo.WebApp/Pages/Admin/Maintenance/Addresses/Index.cshtml
@@ -1,62 +1,0 @@
-﻿@page
-@using Enfo.Domain.Entities.Users
-@model Enfo.WebApp.Pages.Admin.Maintenance.Addresses.Index
-@{
-    ViewData["Title"] = $"Site Maintenance: {Index.ThisOption.PluralName}";
-}
-
-<a asp-page="../Index">← Site Maintenance</a>
-
-<partial name="_AlertPartial" for="Message" />
-
-<h2>
-    @Index.ThisOption.PluralName
-    @if (User.IsInRole(UserRole.SiteMaintenance))
-    {
-        <a asp-page="Add" class="usa-button usa-button-secondary">Add New</a>
-    }
-</h2>
-
-<p>
-    The @Index.ThisOption.PluralName listed below are available for selection when adding or editing Enforcement Orders.
-    Removing one from use makes it unavailable for future Orders but does not affect any existing Orders.
-    The @UserRole.SiteMaintenanceRole.DisplayName role is needed to edit the list.
-</p>
-
-<table class="usa-table-borderless gaepd-table-responsive">
-    <thead>
-    <tr>
-        <th scope="col">Address</th>
-        <th scope="col">Actions</th>
-    </tr>
-    </thead>
-    <tbody>
-    @foreach (var item in Model.Items)
-    {
-        <tr class="@(item.Id == Model.HighlightId ? "gaepd-tablerow-highlight" : item.Active ? "" : "gaepd-tablerow-dimmed")">
-            <th scope="row" data-label="Address">
-                @Html.DisplayFor(m => item.Active, "BoolInactive")
-                @Html.DisplayFor(m => item, "Address")
-            </th>
-            <td class="gaepd-tablecell-nolabel">
-                @if (User.IsInRole(UserRole.SiteMaintenance))
-                {
-                    @if (item.Active)
-                    {
-                        <a asp-page="Edit" asp-route-id="@item.Id" class="gaepd-button-inline-action usa-button">Edit</a>
-                        <form method="post">
-                            <button type="submit" name="id" value="@item.Id" class="gaepd-button-inline-action usa-button-secondary">Remove from use</button>
-                        </form>
-                    }
-                    else
-                    {
-                        <form method="post">
-                            <button type="submit" name="id" value="@item.Id" class="gaepd-button-inline-action usa-button-secondary">Restore</button>
-                        </form>
-                    }
-                }
-            </td>
-        </tr>
-    }
-    </tbody>
-</table>
