+++ src/Enfo.WebApp/Pages/Admin/Maintenance/Contacts/Add.cshtml.cs
@@ -1,7 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Enfo.Domain.Entities.Users;
 using Enfo.Domain.Repositories;
-using Enfo.Domain.Resources.Address;
 using Enfo.Domain.Resources.EpdContact;
 using Enfo.WebApp.Models;
 using Enfo.WebApp.Platform.Extensions;
@@ -9,7 +8,6 @@ using JetBrains.Annotations;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace Enfo.WebApp.Pages.Admin.Maintenance.Contacts
 {
@@ -19,40 +17,25 @@ namespace Enfo.WebApp.Pages.Admin.Maintenance.Contacts
         [BindProperty]
         public EpdContactCreate Item { get; init; }
 
-        public SelectList AddressSelectList { get; private set; }
-
         public static MaintenanceOption ThisOption => MaintenanceOption.EpdContact;
 
         [TempData]
         public int HighlightId { get; [UsedImplicitly] set; }
 
-        private readonly IEpdContactRepository _repository;
-        private readonly IAddressRepository _address;
-
-        public Add(IEpdContactRepository repository, IAddressRepository address) =>
-            (_repository, _address) = (repository, address);
+        [UsedImplicitly]
+        public static void OnGet()
+        {
+            // Method intentionally left empty.
+        }
 
         [UsedImplicitly]
-        public Task OnGetAsync() => PopulateSelectListsAsync();
-
-        [UsedImplicitly]
-        public async Task<IActionResult> OnPostAsync()
+        public async Task<IActionResult> OnPostAsync([FromServices] IEpdContactRepository repository)
         {
             Item.TrimAll();
-
-            if (!ModelState.IsValid)
-            {
-                await PopulateSelectListsAsync();
-                return Page();
-            }
-
-            HighlightId = await _repository.CreateAsync(Item);
+            if (!ModelState.IsValid) return Page();
+            HighlightId = await repository.CreateAsync(Item);
             TempData?.SetDisplayMessage(Context.Success, $"{ThisOption.SingularName} successfully added.");
             return RedirectToPage("Index");
         }
-
-        private async Task PopulateSelectListsAsync() =>
-            AddressSelectList = new SelectList(await _address.ListAsync(), nameof(AddressView.Id),
-                nameof(AddressView.AsLinearString));
     }
 }
