+++ src/Enfo.WebApp/Pages/Admin/Maintenance/Contacts/Index.cshtml
@@ -41,8 +41,7 @@
                 @Html.DisplayFor(m => item, "ContactPerson")
             </th>
             <td data-label="Address">
-                @Html.DisplayFor(m => item.Address.Active, "BoolInactive")
-                @Html.DisplayFor(m => item.Address, "Address")
+                @Html.DisplayFor(m => item, "ContactAddress")
             </td>
             <td data-label="Communication">
                 @Html.DisplayFor(m => item, "ContactCommunication")
