+++ src/Enfo.WebApp/Pages/Admin/Maintenance/Index.cshtml
@@ -13,9 +13,6 @@
     <nav aria-label="Secondary navigation" class="usa-sidenav-list usa-width-one-third">
         <ul class="usa-sidenav-sub_list">
             <li>
-                <a asp-page="Addresses/Index">@MaintenanceOption.Address.PluralName</a>
-            </li>
-            <li>
                 <a asp-page="Contacts/Index">@MaintenanceOption.EpdContact.PluralName</a>
             </li>
             <li>
