+++ src/Enfo.WebApp/Pages/CurrentProposed.cshtml
@@ -64,7 +64,7 @@ else
             {
                 <tr>
                     <th scope="row">@Html.DisplayNameFor(m => enfo.CommentContact)</th>
-                    <td>@Html.DisplayFor(m => enfo.CommentContact, "Contact")</td>
+                    <td>@Html.DisplayFor(m => enfo.CommentContact, "EpdContactView")</td>
                 </tr>
             }
 
@@ -92,7 +92,7 @@ else
                         <p>
                             <em>@Html.DisplayNameFor(m => enfo.HearingContact)</em>
                         </p>
-                        @Html.DisplayFor(m => enfo.HearingContact, "Contact")
+                        @Html.DisplayFor(m => enfo.HearingContact, "EpdContactView")
                     }
                 </tr>
             }
