+++ src/Enfo.WebApp/Pages/Details.cshtml
@@ -57,7 +57,7 @@
         {
             <tr>
                 <th scope="row">@Html.DisplayNameFor(m => m.Item.CommentContact)</th>
-                <td>@Html.DisplayFor(m => m.Item.CommentContact, "Contact")</td>
+                <td>@Html.DisplayFor(m => m.Item.CommentContact, "EpdContactView")</td>
             </tr>
         }
     }
@@ -98,7 +98,7 @@
                 <p>
                     <em>@Html.DisplayNameFor(m => m.Item.HearingContact)</em>
                 </p>
-                @Html.DisplayFor(m => m.Item.HearingContact, "Contact")
+                @Html.DisplayFor(m => m.Item.HearingContact, "EpdContactView")
             }
         </tr>
     }
