+++ src/Enfo.WebApp/Pages/RecentExecuted.cshtml
@@ -78,7 +78,7 @@ else
                         <p>
                             <em>@Html.DisplayNameFor(m => enfo.HearingContact)</em>
                         </p>
-                        @Html.DisplayFor(m => enfo.HearingContact, "Contact")
+                        @Html.DisplayFor(m => enfo.HearingContact, "EpdContactView")
                     }
                 </tr>
             }
