--- src/Enfo.WebApp/Pages/Shared/DisplayTemplates/AddressView.cshtml
@@ -1,14 +1,0 @@
-﻿@model Enfo.Domain.Resources.Address.AddressView
-@if (Model != null)
-{
-    <p class="text-nowrap">
-        @Model.Street
-        @if (!string.IsNullOrWhiteSpace(Model.Street2))
-        {
-            <br />
-            @Model.Street2
-        }
-        <br />
-        @Model.City, @Model.State @Model.PostalCode
-    </p>
-}
