+++ src/Enfo.WebApp/Pages/Shared/DisplayTemplates/ContactAddress.cshtml
@@ -1,0 +1,11 @@
+﻿@model Enfo.Domain.Resources.EpdContact.EpdContactView
+<p class="text-nowrap">
+    @Model.Street
+    @if (!string.IsNullOrWhiteSpace(Model.Street2))
+    {
+        <br />
+        @Model.Street2
+    }
+    <br />
+    @Model.City, @Model.State @Model.PostalCode
+</p>
