+++ src/Enfo.WebApp/Platform/DevHelpers/TempData.cs
@@ -7,6 +7,7 @@ using Enfo.Domain.Utils;
 using Enfo.Infrastructure.Contexts;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
+// ReSharper disable StringLiteralTypo
 
 namespace Enfo.WebApp.Platform.DevHelpers
 {
@@ -15,14 +16,6 @@ namespace Enfo.WebApp.Platform.DevHelpers
         public static async Task SeedTempDataAsync([NotNull] this EnfoDbContext c, CancellationToken t)
         {
             await c.Database.OpenConnectionAsync(t);
-
-            if (!await c.Addresses.AnyAsync(t))
-            {
-                await c.Database.ExecuteSqlRawAsync($"SET IDENTITY_INSERT dbo.{nameof(c.Addresses)} ON", t);
-                await c.Addresses.AddRangeAsync(GetAddresses, t);
-                await c.SaveChangesAsync(t);
-                await c.Database.ExecuteSqlRawAsync($"SET IDENTITY_INSERT dbo.{nameof(c.Addresses)} OFF", t);
-            }
 
             if (!await c.LegalAuthorities.AnyAsync(t))
             {
@@ -51,38 +44,6 @@ namespace Enfo.WebApp.Platform.DevHelpers
             await c.Database.CloseConnectionAsync();
         }
 
-        private static readonly IEnumerable<Address> GetAddresses = new List<Address>
-        {
-            new()
-            {
-                Id = 2000,
-                Active = true,
-                City = "Atlanta",
-                PostalCode = "30354",
-                State = "GA",
-                Street = "4244 International Parkway",
-                Street2 = "Suite 120",
-            },
-            new()
-            {
-                Id = 2001,
-                Active = false,
-                City = "Atlanta",
-                PostalCode = "30354",
-                State = "GA",
-                Street = "000 Obsolete Address",
-            },
-            new()
-            {
-                Id = 2002,
-                Active = true,
-                City = "Atlanta",
-                PostalCode = "30000",
-                State = "GA",
-                Street = "123 New Street",
-            },
-        };
-
         private static readonly IEnumerable<LegalAuthority> GetLegalAuthorities = new List<LegalAuthority>
         {
             new() {Id = 1, Active = true, AuthorityName = "Air Quality Act",},
@@ -96,34 +57,45 @@ namespace Enfo.WebApp.Platform.DevHelpers
             {
                 Id = 2000,
                 Active = true,
-                AddressId = 2000,
                 ContactName = "A. Jones",
                 Email = "example@example.com",
                 Organization = "Environmental Protection Division",
                 Telephone = "555-1212",
                 Title = "Chief, Air Protection Branch",
+                City = "Atlanta",
+                PostalCode = "30354",
+                State = "GA",
+                Street = "4244 International Parkway",
+                Street2 = "Suite 120",
             },
             new()
             {
                 Id = 2001,
                 Active = false,
-                AddressId = 2001,
                 ContactName = "B. Smith",
                 Email = null,
                 Organization = "Environmental Protection Division",
                 Telephone = null,
                 Title = "Chief, Land Protection Branch",
+                City = "Atlanta",
+                PostalCode = "30354",
+                State = "GA",
+                Street = "4244 International Parkway",
+                Street2 = "Suite 120",
             },
             new()
             {
                 Id = 2002,
                 Active = true,
-                AddressId = 2001,
                 ContactName = "B. Smith",
                 Email = null,
                 Organization = "Environmental Protection Division",
                 Telephone = null,
                 Title = "Chief, Land Protection Branch",
+                City = "Atlanta",
+                PostalCode = "30000",
+                State = "GA",
+                Street = "123 New Street",
             },
         };
 
@@ -506,11 +478,12 @@ namespace Enfo.WebApp.Platform.DevHelpers
                 ExecutedDate = DateTime.Today.AddDays(-14),
                 ExecutedOrderPostedDate = DateUtils.MostRecentMonday(),
                 FacilityName = "bcd22-" + Guid.NewGuid(),
-                HearingCommentPeriodClosesDate = null,
-                HearingDate = null,
-                HearingLocation = "",
+                HearingCommentPeriodClosesDate = new DateTime(2012, 11, 21),
+                HearingContactId = 2000,
+                HearingDate = DateTime.Today.AddDays(14),
+                HearingLocation = "bcd22-" + Guid.NewGuid(),
                 IsExecutedOrder = true,
-                IsHearingScheduled = false,
+                IsHearingScheduled = true,
                 IsProposedOrder = false,
                 LegalAuthorityId = 2,
                 OrderNumber = "EPD-WP-0022",
